Create modules for book, localstorage and UI
